requirements.txt
requests
beautifulsoup4
SQLAlchemy
lxml
pandas
numpy
python-dotenv